[#351] addShadowWithOpaqueBackground() 자료형 이슈

+++ Mohaeng/Mohaeng/Sources/Extensions/UIView+Extension.swift
@@ -80,7 +80,7 @@ extension UIView {
         self.layer.masksToBounds = false
     }
     
-    func addShadowWithOpaqueBackground(opacity: CGFloat, radius: CGFloat) {
+    func addShadowWithOpaqueBackground(opacity: Float, radius: CGFloat) {
         self.layer.shadowOpacity = opacity
         self.layer.shadowOffset = CGSize(width: 0, height: 0)
         self.layer.shadowRadius = radius
